+++ examples/advanced.py
@@ -2,6 +2,13 @@
 
 import os
 
+try:
+    from dotenv import load_dotenv
+
+    load_dotenv()
+except ImportError:
+    pass
+
 from zerotrue import ZeroTrue
 
 client = ZeroTrue(