+++ zerotrue/__init__.py
@@ -2,14 +2,9 @@
 
 from zerotrue.async_client import AsyncZeroTrue
 from zerotrue.client import ZeroTrue
-from zerotrue.exceptions import (
-    APIError,
-    AuthenticationError,
-    RateLimitError,
-    ValidationError,
-)
+from zerotrue.exceptions import APIError, AuthenticationError, RateLimitError, ValidationError
 
-__version__ = "1.1.0"
+__version__ = "1.2.0"
 __all__ = [
     "ZeroTrue",
     "AsyncZeroTrue",